segregate IC and custom resource fixtures
